requirements.txt
flask
python-dotenv
gunicorn
alpaca-trade-api
flask_sqlalchemy
flask_migrate
pandas
pandas_datareader
numpy
datetime
finvizfinance
progress
xlsxwriter
psycopg2-binary